showdropdown page was added
