+++ lib/pages/user_approval_detail.dart
@@ -182,22 +182,18 @@ class _PageContentState extends State<PageContent> {
                                   ),
                                 ),
                                 onPressed: () {
-                                  /*if(widget.userApproval.type == null){
+                                  if (widget.userApproval.type == null) {
                                     final accounts = widget.parseMessages?.call() ?? [];
-                                    _showApprovalTableDialog(accounts);*/
-                                    if (widget.userApproval.type == null) {
-                                        final accounts = widget.parseMessages?.call() ?? [];
-                                        // Navigate to the new dialog page and pass the accounts list
-                                        Navigator.push(
-                                            context,
-                                            MaterialPageRoute(
-                                              builder: (context) => ApprovalTableDialogPage(
-                                                accounts: accounts,// Pass the list of accounts
-                                                showDropdown: _showDropdown, // Pass the _showDropdown function
-                                                userApproval: widget.userApproval, userPayload: [],
-                                              ),
-                                            ),
-                                          );
+                                    // Navigate to the new dialog page and pass the accounts list
+                                    Navigator.push(
+                                      context,
+                                      MaterialPageRoute(
+                                        builder: (context) => ApprovalTableDialogPage(
+                                          accounts: accounts,
+                                          userApproval: widget.userApproval, userPayload: [],
+                                        ),
+                                      ),
+                                    );
                                   } else if (widget.userApproval.type == "deactivate" || widget.userApproval.type == "activate"){
                                     context.read<MessageModel>().approvalActRequest(widget.userApproval);
                                   }
@@ -291,433 +287,6 @@ class _PageContentState extends State<PageContent> {
     }
   }
 
-
-  Future<void> _showApprovalTableDialog(List<dynamic> accounts) async {
-    showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: Text('Review User Details'),
-          content: Container(
-            width: MediaQuery.of(context).size.width * 0.9,
-            constraints: BoxConstraints(
-                maxHeight: MediaQuery.of(context).size.height * 0.6),
-            child: SingleChildScrollView(
-              child: ListView.builder(
-                shrinkWrap: true,
-                physics: ClampingScrollPhysics(),
-                itemCount: accounts.length,
-                itemBuilder: (context, index) {
-                  final account = accounts[index];
-                  final payloadStatus = account['payloadStatus'];
-                  Color? rowColor;
-
-                  if (payloadStatus == 'CREATED') {
-                    rowColor = Colors.green.withOpacity(0.2);
-                  } else if (payloadStatus == 'REJECTED') {
-                    rowColor = Colors.red.withOpacity(0.2);
-                  } else {
-                    rowColor = Colors.transparent;
-                  }
-
-                  return Container(
-                    color: rowColor,
-                    child: ExpansionTile(
-                      title: Text(
-                        '${account['SN'] ?? ''} : ${account['Names'] ?? ''}',
-                        style: TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                      children: [
-                        SingleChildScrollView(
-                          child: Column(
-                            children: [
-                              ListTile(
-                                title: Text('Email: ${account['Email'] ?? 'N/A'}'),
-                              ),
-                              ListTile(
-                                title: Text('Phone Number: ${account['Phone Number'] ?? 'N/A'}'),
-                              ),
-                              ListTile(
-                                title: Text('Entry Access Level: ${account['Entry Access Level'] ?? 'N/A'}'),
-                              ),
-                              ListTile(
-                                title: Text('Report Access Level: ${account['Report Access Level'] ?? 'N/A'}'),
-                              ),
-                              ListTile(
-                                title: Text(
-                                  'Action:',
-                                  style: TextStyle(fontWeight: FontWeight.bold),
-                                ),
-                                subtitle: payloadStatus == 'CREATED'
-                                    ? Text(
-                                  'Created',
-                                  style: TextStyle(color: const Color.fromARGB(255, 3, 56, 4)),
-                                )
-                                    : payloadStatus == 'REJECTED'
-                                    ? Text(
-                                  'Rejected',
-                                  style: TextStyle(color: const Color.fromARGB(255, 91, 13, 7)),
-                                )
-                                    : ElevatedButton(
-                                  onPressed: () {
-                                    final nameParts = (account['Names'] ?? '').split(' ');
-                                    final firstName = nameParts.isNotEmpty ? nameParts[0] : '';
-                                    final lastName = nameParts.length > 1 ? nameParts[1] : '';
-                                    _showDropdown(context, firstName, lastName, widget.userApproval);
-                                  },
-
-                                  child: Text('Select'),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  );
-                },
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: Text('Cancel'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-
-  void _showPotentialDuplicatesPopup(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Potential Duplicates'),
-          content: Container(
-            width: MediaQuery.of(context).size.width * 0.9,
-            constraints: BoxConstraints(
-                maxHeight: MediaQuery.of(context).size.height * 0.4),
-            child: Column(
-              children: [
-                Expanded(
-                  child: ListView.builder(
-                    itemCount: _potentialDuplicates.length,
-                    itemBuilder: (context, index) {
-                      final user = _potentialDuplicates[index];
-
-                      return ExpansionTile(
-                        title: Text('${(index + 1).toString()} : ${user['firstName'] ?? 'Unknown'} ${user['surname'] ?? 'Unknown'}'),
-                        children: <Widget>[
-                          ListTile(
-                            title: Text('Phone Number:'),
-                            subtitle: Text(user['phoneNumber'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Email:'),
-                            subtitle: Text(user['email'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Data Entry:'),
-                            subtitle: Text(user['dataEntry'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Reports:'),
-                            subtitle: Text(user['reports'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Username:'),
-                            subtitle: Text(user['username'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Last Login:'),
-                            subtitle: Text(user['lastLogin'] ?? 'N/A'),
-                          ),
-                          ListTile(
-                            title: Text('Password Last Updated:'),
-                            subtitle: Text(user['passwordLastUpdated'] ?? 'N/A'),
-                          ),
-                        ],
-                      );
-                    },
-                  ),
-                ),
-              ],
-            ),
-          ),
-          actions: <Widget>[
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: Text('Close'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future <void> _showDropdown(BuildContext context, String firstName, String lastName, UserModel userApproval) async {
-    final usernameController = TextEditingController();
-    String proposedUsername = _generateProposedUsername(firstName, lastName);
-    usernameController.text = proposedUsername;
-    _showDuplicatesButton = false;
-    var SelectedPayload = userApproval.userPayload!.firstWhere(
-          (payloadu) => '${payloadu.firstName} ${payloadu.surname}' == '$firstName $lastName',);
-
-    Future <bool> _checkExistingUsername(String username) async {
-      EasyLoading.show(
-        status: 'Checking for Existing Username...🔄',
-        maskType: EasyLoadingMaskType.black,
-      );
-      final response = await d2repository.httpClient.get(
-        'users?filter=userCredentials.username:eq:$username&fields=id',
-      );
-
-      if (response.statusCode == 200) {
-        List<dynamic> data = response.body['users']??[0]??['id'].toList();
-        if (data.isEmpty) {
-          EasyLoading.showSuccess('No Existing Username found');
-          return false;
-        } else {
-          EasyLoading.showSuccess('Existing Username found... Please Change the Username');
-          proposedUsername = _suggestAlternativeUsername(proposedUsername, firstName, username);
-          usernameController.text = proposedUsername;
-          return true;
-        }
-      } else {
-
-        _showMessage(context, "Error checking existing username.");
-        return true;
-      }
-    }
-
-    Future<bool> _checkDuplicate(UserModel userApproval, Userpayload SelectedPayload) async {
-      EasyLoading.show(
-        status: 'Performing Check...🔄 ',
-        maskType: EasyLoadingMaskType.black,
-      );
-      if (userApproval.userPayload != null && userApproval.userPayload!.isNotEmpty) {
-        String? email = SelectedPayload.email;
-        String? phoneNumber = SelectedPayload.phoneNumber;
-        String? userApprovalId = userApproval.id;
-
-        if (email == null || phoneNumber == null || userApprovalId == null) {
-          print("Email, PhoneNumber, or UserApprovalId is missing.");
-          EasyLoading.showError('Missing Data.');
-          return false;
-        }
-
-        try {
-          print('Phone number query: $phoneNumber');
-          print('Email query: $email');
-
-          final emailResponse = await d2repository.httpClient.get(
-              'users.json?filter=email:eq:$email&fields=id,firstName,surname,email,phoneNumber,userCredentials,disabled,createdBy[name],lastUpdatedBy[name],username,lastLogin,passwordLastUpdated,organisationUnits[id,name],dataViewOrganisationUnits[id,name]'
-          );
-
-          final phoneResponse = await d2repository.httpClient.get(
-              'users.json?filter=phoneNumber:eq:$phoneNumber&fields=id,firstName,surname,email,phoneNumber,userCredentials,disabled,createdBy[name],lastUpdatedBy[name],username,lastLogin,passwordLastUpdated,organisationUnits[id,name],dataViewOrganisationUnits[id,name]'
-          );
-
-          dynamic emailResponseBody;
-          if (emailResponse.body is String) {
-            emailResponseBody = jsonDecode(emailResponse.body);
-          } else {
-            emailResponseBody = emailResponse.body;
-          }
-
-          dynamic phoneResponseBody;
-          if (phoneResponse.body is String) {
-            phoneResponseBody = jsonDecode(phoneResponse.body);
-          } else {
-            phoneResponseBody = phoneResponse.body;
-          }
-
-          if (emailResponseBody is! Map<String, dynamic> || phoneResponseBody is! Map<String, dynamic>) {
-            print('Error: Response bodies are not maps');
-            EasyLoading.showError('Invalid Response Format');
-            return false;
-          }
-
-          if ((emailResponse.statusCode == 200 || emailResponse.statusCode == 304) &&
-              (phoneResponse.statusCode == 200 || phoneResponse.statusCode == 304)) {
-
-            final totalEmailUsers = emailResponseBody['pager']['total'];
-            final totalPhoneUsers = phoneResponseBody['pager']['total'];
-
-            print('Total Email Users Found: $totalEmailUsers');
-            print('Total Phone Users Found: $totalPhoneUsers');
-
-            List<dynamic> emailUsers = emailResponseBody['users'] ?? [];
-            List<dynamic> phoneUsers = phoneResponseBody['users'] ?? [];
-
-            print('Email Users: $emailUsers');
-            print('Phone Users: $phoneUsers');
-
-            bool isDuplicateEmail = emailUsers.any((user) => (user['id'] as String?) != userApprovalId);
-            bool isDuplicatePhone = phoneUsers.any((user) => (user['id'] as String?) != userApprovalId);
-
-            if (isDuplicateEmail || isDuplicatePhone) {
-              _potentialDuplicates = [...emailUsers, ...phoneUsers];
-
-              EasyLoading.showSuccess('Succeeded check!, duplicates were found');
-              return true;
-            }
-
-            EasyLoading.showSuccess('Succeeded check!, No duplicates found!');
-            return false;
-          } else {
-            print('Failed to fetch user data for email or phone number');
-            EasyLoading.showError('Failed to fetch user data');
-            return false;
-          }
-        } catch (e, stackTrace) {
-          print('An error occurred: $e');
-          print('Stack trace: $stackTrace');
-          EasyLoading.showError('Error during duplicate check');
-          return false;
-        } finally {
-          EasyLoading.dismiss();
-        }
-      } else {
-        print("No payload data available.");
-        EasyLoading.showError('No payload data available.');
-        return false;
-      }
-    }
-
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Username Details for $firstName $lastName'),
-          content: Container(
-            width: MediaQuery.of(context).size.width * 0.9,
-            child: SingleChildScrollView(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  TextField(
-                    controller: usernameController,
-                    decoration: InputDecoration(labelText: 'Username (Ensure five or more Characters)'),
-                  ),
-                  SizedBox(height: 20),
-                  ElevatedButton(
-                    onPressed: () async {
-                      bool hasDuplicates = await _checkDuplicate(userApproval, SelectedPayload);
-                      setState(() {
-                        _showDuplicatesButton = hasDuplicates;
-                      });
-                    },
-                    child: Text('Check Duplicate'),
-                  ),
-                  if (_showDuplicatesButton)
-                    ElevatedButton(
-                      onPressed: () => _showPotentialDuplicatesPopup(context),
-                      child: Text('View Potential Duplicates'),
-                    ),
-                  SizedBox(height: 30),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      ElevatedButton(
-                        onPressed: () {
-                          showDataAlert(context, SelectedPayload, false);
-                        },
-                        child: Text('Reject'),
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.red,
-                        ),
-                      ),
-                      ElevatedButton(
-                        onPressed: () async {
-                          bool usernameExists = await _checkExistingUsername(usernameController.text);
-                          if (!usernameExists){
-                            _confirmUsername(context, usernameController.text, SelectedPayload);
-                          }
-                        },
-                        child: Text('Confirm'),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  String _generateProposedUsername(String firstName, String lastName) {
-    return firstName[0].toLowerCase() + lastName.toLowerCase();
-  }
-
-  String _suggestAlternativeUsername(String baseUsername, String firstName, String existingUsernames){
-    String suggestion = baseUsername;
-
-    for (int i = 1; i <= firstName.length; i++) {
-      suggestion = baseUsername + firstName.substring(0, i).toLowerCase();
-      if (!existingUsernames.contains(suggestion)) {
-        return suggestion;
-      }
-    }
-
-    int number = 1;
-    while (existingUsernames.contains(suggestion)) {
-      suggestion = baseUsername + number.toString();
-      number++;
-    }
-
-    return suggestion;
-  }
-
-  Future<void> _confirmUsername(BuildContext context, String username, Userpayload SelectedPayload) async {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Confirmation'),
-          content: Text('Are you sure you want to create a user with username $username?'),
-          actions: [
-            TextButton(
-              onPressed: () {
-                _createUser(username, SelectedPayload);
-                Navigator.of(context).pop();
-              },
-              child: Text('Yes'),
-            ),
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: Text('No'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future <void> _createUser(String username, Userpayload SelectedPayload, {bool isAccept = true}) async {
-    if (widget.userPayload != null && widget.userPayload!.isNotEmpty) {
-      SelectedPayload.username = username;
-      SelectedPayload.password = "Hmis@2024";
-      print('User created with username: ${SelectedPayload.username}');
-    }
-    await _loading(isAccept, SelectedPayload);
-    Navigator.of(context).pop();
-  }
-
-  void _showMessage(BuildContext context, String message) {
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(message)));
-  }
-
   Future  <void> showDataAlert(BuildContext context, Userpayload SelectedPayload, bool isRejectAll, {bool isAccept = false}) async {
     showDialog(
       context: context,
@@ -794,7 +363,9 @@ class _PageContentState extends State<PageContent> {
       },
     );
   }
+
 }
+
 
 String extractPlainText(String htmlContent) {
   final document = html_parser.parse(htmlContent);
